Created 'conway_classes.js'. Separated HTML and JS.
The 'conway_classes.js' contains a ConwayBoard class and a
ConwayCell class.  The ConwayBoard class accesses a HTML5 canvas
element and renders its cells using a 'render' function.

The 'index.html' file now includes the JavaScript classes inside
the head element.  The only JavaScript that is explicitly embedded
inside the 'index.html' is an instantiation of a ConwayBoard and
a call to its render method.