Changed way that canvas offset is determined.
Previously, the 'offsetLeft' and 'offsetTop' properties of the
canvas element were being used to retrieve the position of the
top-left corner of the canvas. The values of these properties,
however, are determined relative to the parent element. To support
nesting of the canvas within several HTML elements, the
'<element>.getBoundingClientRect()' method was used intead to
determine the canvas' offsets.  See the 'processClick' function
for this change.

Minor changes were made elsewhere.  Need to refine the code.